requirements.txt
Django>=4.2.0
paho-mqtt>=1.6.1
scikit-learn>=1.0.2
numpy>=1.21.0
joblib>=1.1.0 